refactor: ExpressionParser 추상 클래스화
