+++ src/main/kotlin/com/nextstep/second/calculator/README.md
@@ -17,7 +17,7 @@
 * [X] ExpressionParser : input 을 받아서 List<Int> 로 숫자들을 파싱해준다
   * `String` 타입의 Input 을 받아서 regex 를 이용하여 List<Int> 로 분리해줄 수 있음
   * `() -> Unit` 타입의 validFunction 을 입력받아서 validation 을 진행 -> 일단 private 내장 함수로 구현
-* [ ] CustomExpressionParser : input 을 받아서 List<Int> 로 숫자를 파싱해준다
+* [X] CustomExpressionParser : input 을 받아서 List<Int> 로 숫자를 파싱해준다
   * `String` 타입의 Input 을 받아서 regex 를 이용하여 List<Int> 로 분리해줄 수 있음
   * `//[.]\n` 에서 기준선을 잡아서 이 후 문자열을 파싱하여 List<Int> 로 분리해줄 수 있음
   * `() -> Unit` 타입의 validFunction 을 입력받아서 validation 을 진행
